+++ tests/unit/Deserializers/ItemDeserializerTest.php
@@ -2,9 +2,14 @@
 
 namespace Tests\Wikibase\DataModel\Deserializers;
 
+use Wikibase\DataModel\Claim\Claim;
 use Wikibase\DataModel\Deserializers\ItemDeserializer;
 use Wikibase\DataModel\Entity\Item;
+use Wikibase\DataModel\Entity\ItemId;
 use Wikibase\DataModel\SiteLink;
+use Wikibase\DataModel\Snak\PropertyNoValueSnak;
+use Wikibase\DataModel\Statement\Statement;
+use Wikibase\DataModel\Statement\StatementList;
 use Wikibase\DataModel\Term\Fingerprint;
 
 /**
@@ -17,15 +22,36 @@ class ItemDeserializerTest extends DeserializerBaseTest {
 
 	public function buildDeserializer() {
 		$entityIdDeserializerMock = $this->getMock( '\Deserializers\Deserializer' );
+		$entityIdDeserializerMock->expects( $this->any() )
+			->method( 'deserialize' )
+			->will( $this->returnValue( new ItemId( 'Q42' ) ) );
 
 		$fingerprintDeserializerMock = $this->getMock( '\Deserializers\Deserializer' );
 		$fingerprintDeserializerMock->expects( $this->any() )
 			->method( 'deserialize' )
 			->will( $this->returnValue( new Fingerprint() ) );
 
-		$statementListDeserializerMock = $this->getMock( '\Deserializers\Deserializer' );
-		$siteLinkDeserializerMock = $this->getMock( '\Deserializers\Deserializer' );
+		$statement = new Statement( new Claim( new PropertyNoValueSnak( 42 ) ) );
+		$statement->setGuid( 'test' );
 
+		$statementListDeserializerMock = $this->getMock( '\Deserializers\Deserializer' );
+		$statementListDeserializerMock->expects( $this->any() )
+			->method( 'deserialize' )
+			->with( $this->equalTo( array(
+				'P42' => array(
+					array(
+						'mainsnak' => array(
+							'snaktype' => 'novalue',
+							'property' => 'P42'
+						),
+						'type' => 'statement',
+						'rank' => 'normal'
+					)
+				)
+			) ) )
+			->will( $this->returnValue( new StatementList( array( $statement ) ) ) );
+
+		$siteLinkDeserializerMock = $this->getMock( '\Deserializers\Deserializer' );
 		$siteLinkDeserializerMock->expects( $this->any() )
 			->method( 'deserialize' )
 			->with( $this->equalTo( array(
@@ -72,6 +98,36 @@ class ItemDeserializerTest extends DeserializerBaseTest {
 					'type' => 'item'
 				)
 			),
+		);
+
+		$item = new Item( new ItemId( 'Q42' ) );
+		$provider[] = array(
+			$item,
+			array(
+				'type' => 'item',
+				'id' => 'Q42'
+			)
+		);
+
+		$item = new Item();
+		$item->getStatements()->addNewStatement( new PropertyNoValueSnak( 42 ), null, null, 'test' );
+		$provider[] = array(
+			$item,
+			array(
+				'type' => 'item',
+				'claims' => array(
+					'P42' => array(
+						array(
+							'mainsnak' => array(
+								'snaktype' => 'novalue',
+								'property' => 'P42'
+							),
+							'type' => 'statement',
+							'rank' => 'normal'
+						)
+					)
+				)
+			)
 		);
 
 		$item = new Item();
