+++ tests/unit/Deserializers/PropertyDeserializerTest.php
@@ -5,6 +5,7 @@ namespace Tests\Wikibase\DataModel\Deserializers;
 use Wikibase\DataModel\Claim\Claim;
 use Wikibase\DataModel\Deserializers\PropertyDeserializer;
 use Wikibase\DataModel\Entity\Property;
+use Wikibase\DataModel\Entity\PropertyId;
 use Wikibase\DataModel\Snak\PropertyNoValueSnak;
 use Wikibase\DataModel\Statement\Statement;
 use Wikibase\DataModel\Statement\StatementList;
@@ -20,6 +21,9 @@ class PropertyDeserializerTest extends DeserializerBaseTest {
 
 	public function buildDeserializer() {
 		$entityIdDeserializerMock = $this->getMock( '\Deserializers\Deserializer' );
+		$entityIdDeserializerMock->expects( $this->any() )
+			->method( 'deserialize' )
+			->will( $this->returnValue( new PropertyId( 'P42' ) ) );
 
 		$fingerprintDeserializerMock = $this->getMock( '\Deserializers\Deserializer' );
 		$fingerprintDeserializerMock->expects( $this->any() )
@@ -86,6 +90,38 @@ class PropertyDeserializerTest extends DeserializerBaseTest {
 					'datatype' => 'string'
 				)
 			),
+		);
+
+		$property = new Property( new PropertyId( 'P42' ), null, '' );
+		$provider[] = array(
+			$property,
+			array(
+				'type' => 'property',
+				'datatype' => '',
+				'id' => 'P42'
+			)
+		);
+
+		$property = Property::newFromType( '' );
+		$property->getStatements()->addNewStatement( new PropertyNoValueSnak( 42 ), null, null, 'test' );
+		$provider[] = array(
+			$property,
+			array(
+				'type' => 'property',
+				'datatype' => '',
+				'claims' => array(
+					'P42' => array(
+						array(
+							'mainsnak' => array(
+								'snaktype' => 'novalue',
+								'property' => 'P42'
+							),
+							'type' => 'statement',
+							'rank' => 'normal'
+						)
+					)
+				)
+			)
 		);
 
 		$property = Property::newFromType( '' );
